Made the overriding of theme formatting more "sane".
If the theme nick or timestamp is overridden, then it overrides ANY
setting.  That is, even if the Theme defines its own nick and timestamp
format, having the option for custom timestamp overrides this setting.

If the theme doesn't define it's own nick / timestamp format, then when
the checkbox is _not_ checked it uses the default settings.

In the "Style Preferences" dialog, when someone leaves the nick or time
format combo boxes are left blank, then the default value is displayed
(setting in the properties of the combo box).