Strip mode prefix chars for nick completion.
Nick names with +o or +v get a "@" or "+" character prefixed
in the user list. Nick tab-completion requires the extra char
to be typed for the completion to match, resulting in IRC-
unusual nick+mode-char conversations.

This patch removes a leading @ or + from each nick in the
user list prior to matching for nick completion.

I don't know if this is the best way to do this, but it works :)