Fixes #178 Fixes #191
This commit provide staging for advancing the behavior of the input text field of Textual.

Notable changes:
- Support for up and down arrows to browse the multiline text field instead of only being used for history.
- Internal drawing is now based off the total number of lines in text field. This is a very important change. No more weird heights for the use of emoji icons and other strange symbols. Those used to distort the text field because the old method measured the height of the text. We now use a static value for the height of each line.

These changes are considered still buggy in themselves, but they are designed to fix some common issues with Textual. These changes are also related to open input history bugs which are being worked on at the same time.