Sleep mode checkbox.
There has been a lot of feedback regarding how Textual handles sleep mode.
The default behavior was to always disconnect all clients when sleep mode is
activated. In many cases, such as an unplugged MacBook, OS X will turn the
network off to save battery. That is why we disconnected all clients so the user
would not mistaken the disconnect on sleep mode as a timeout.

This is also done on computers plugged in that do not turn off the network. Now
a checkbox is included in server properties as to whether the client should be
disconnected and reconnected on sleep mode. On by default to replicate old
behavior.