Show lock icon next to encrypted messages.
The lock icon is opt-in by themes and the template has to be placed
in the directory of the theme itself in order to support it. This is done
because Textual does not know whether a theme needs a lock icon
for a dark or light background.

Two icons are supplied by default for themes to use as their lock. They
can also include their own if necessary.