New command for developer mode: "FAKEINDATA"
Generates a random number of bots (1-10). These bots join the channel in which
the command is being executed and picks from a list of IRC commands to execute.
The pick is random. The commands executed are JOIN, PART, QUIT, NICK, PRIVMSG,
ACTION, NOTICE, and TOPIC. These commands are never actually sent to the server.
They are pushed internally to Textual to replicate a chatroom with a little bit of activity.

This command is designed to allow theme developers to test various line types without
having to join an actual channel and waiting for them to occur. It is recommended to
run this command on a disconnected server. Results on a server that is already connected
may result in irregular activity by Textual.