Updated build configuration.
Configuration paramaters such as framework search path, header search
path, build product location, and several other folders are now passed
to the Build Framework and Build Extensions scripts instead of being
defined in their configuration files. This is done so that only the
main Textual configuration needs to be updated for the passing of this
information.

Additionally, the actual configuration of each target has been updated
to share the same values. This includes the main Textual executable,
the frameworks, and the extensions.

Build results for everything is now stored in a .tmp folder in the root
directory. The Textual executable result is still located in the
"BuildResults" folder.