Work In Progress: View Based Server List
Do not compile this. I am only committing to back up my work remotely incase of a lack of Internet.
This is a work-in-progress of converting the serve list of Textual (and eventually the user list) to a
view based layout. This new design uses a lot of custom drawing to get done what it does and is
not ready for real use. It is not optimized right now.