CTCP Version reply changes. See the details of this commit for exact details.
This commit introduces a new format for CTCP version replies for Textual. The changes also apply to the "/myversion" command.
The changes were done to get rid of most of the bulk part of the version replies which are not relevant to almost all users.

New CTCP Version Reply: Textual <version major> (Flavor: <version code name>)
Old CTCP Version Reply: Textual <version major> (Git Reference: <build git ref>) (Flavor: <version code name>)

The format of the MYVERSION command has also been changed. It uses the new reply except for some minor differences.
Instead of flavor, MYVERSION shows the build number which is the total number of commits for the Textual repo. As this
number is always increasing, this can be used as a reference to a specific one. Appending "-d" to the MYVERSION command
now outputs the full version information. This includes the flavor and the git reference.