User interface improvements.
The welcome sheet now makes getting to the support channel much more easier.
First thing the user see is now "First, we need to add a server or you can chose to
connect to our support channel." instead of only the first half of that sentence. The
"support channel" part is a link which Textual will intercept and handle the connection.

This commit also includes a patch for createClient:reload: to select the newly created
client if nothing existed before it.