Dynamic icon for alerts.
If Growl is running, then the icon for the Alerts tab of Preferences will be the Growl logo.
If it is not running and we are on Mountain Lion, then we use the Notification Center icon.
This is a subtle way of telling where Textual (or the Growl framework) will send the notifications.