Use model name instead of model identifier.
For the SYSINFO command, when our MacintoshModels property list does not contain a specific model identifier becase
we did not have time to update it when a new Mac is released… we now use the actual model name instead of the model
identifier that we found. For example, instead of showing the user "iMac11,1" we only show them "iMac" or for "MacBookAir"
we show "Macbook Air" — much cleaner and makes us look less stupid. :-)