Allow exclude words in full word matching mode.
The policy to not allow exclude words in full word matching mode was inherited from Limechat. Never modified the policy because there was never a demand. Now that keywords can be both global and server specific it makes more sense to remove this policy and allow exclusion regardless of the match type.