Redesign of Alerts tab of Preferences.
This was done to make the user more aware that the settings presented on the pane are those of the selected alert.
This commit removes the option to disable notifications when Textual is in focus. Instead, this setting is always used.
Notifications will not be posted while Textual is in focus.