Forced a standard format for log files.
This makes it so user defined nickname formatting and timestamps are
not used in log files. Resolves an issue related to styles defining
blank nickname/timestamp formats resulting in weird logginb behavior.
Forcing a standard format for regular messages also makes log files
easier to be pushed through a parser.