Add support for minimum width formatting in nick
Adds support for specifying a minimum width for nick display. Both the nick
(`%n`) and mark (`%@`) attributes can have specify a minimum width by adding
an integer describing the width following the percent character (`%`). If
a literal percent is desired in the formatted nick it must be escaped in the
format string by doubling (eg. `%%).

Example:
- format: `<%-9n>`
- nick: `bd808`
- result: `<    bd808>`.