Change type= to ltype= in styles.  (see description)
In response to bug #119000 filed with the WebKit project as well as rdar://problem/14643850 filed with Apple; the type= selector must be renamed
as it creates an annoying bug in Textual on Mac OS Mavericks. I filed both those bug reports months ago and still nothing has came of them so it
is time to force change. The type= selector will still remain part of the default templates, but the transition to ltype= is necessary.