Add support for importing .TxtlPlugin files.
Concept is simple. Textual is given URL, it copies it to its sandbox. If a plugin already exists with that name, it moves it to trash, then copies new one.
Reloading plugins is a very delicate process so we aren’t even going to try and reload it for the plugin.