Add delete client logic. (not tested)
Textual maintains a list of all deleted clients unique identifiers which is synced between all iCloud enabled devices. When a client is deleted by the end user, its unique identifier is added to this list. Upon deletion, if the client was configured to be excluded from iCloud syncing, then its actually not added to the deleted client list.

If a clients unique identifier exists in the deleted client list, but is configured locally to not be synced to iCloud, then it is not deleted upon sync of the list. If that local client is every updated to sync back to iCloud, then we delete its unique identifier from the deleted clients list so other synced devices do not try to delete it upon addition.