Rewrite some internals of the plugin manager. May break some things.
* Plugin unloads are no longer allowed to be performed by any means.
The process of unloading and reloading has always been unstable because
of the nature of bundles so it is best not to include it at all so that
the bad behavior is not encouraged. The protocol selector
“pluginUnloadedFromMemory” is still called when Textual is shutting
down. This change also makes it possible for plugins to register for
defaults without that pesky crash during reloads.
* The internal process of loading plugins now occur on a worker thread
like the rest of the plugin selectors. It is no longer the main thread.
This may break some plugins.
* The protocol selector “processInlineMediaContentURL:” is no longer
half-assed. It requires a valid URL to be returned or it goes to the
next plugin.

The documentation has been updated to reflect these changes.