Auto validate SSL certificate on reconnects.
If we are doing a reconnect it is assumed that the connection already existed and was accepted by the end user as a certificate that they want to use.
If they do not explicitly define it as a disconnect and its just Textual reconnecting after a network drop, then we auto validate the SSL certificate on behalf of the user.

The idea behind this is to allow reconnects during network drops without user being present to accept the certificate again. Running the prompt for validation will hold the main loop and if the user leaves their Mac unattended for several hours, we do not want the prompt just hanging there and a reconnect to never occur.