Fixed issue related to duplicated servers with a password not launching server properties for the duplicate.
The duplicated server has *IsSet for a keychain item so when the server properties dialog tries to set the
string value of that keychain item to the text field, it sets a nil value resulting in the dialog faulting.

Now, while duplicating, we set the existing keychain entries to an empty string to reset the state.