Present the SSL certificate trust panel as a sheet instead of modal to stop it from holding main thread.
Without an exception for reconnects, holding the main thread while waiting for user verification would be a little screwy.

Instead, a sheet is presented which does not hold the thread and waits for user acceptance. If they wish to perform reconnects without the dialog, then they can select to always trust that particular certificate.

This commit is kind of hackish because it chains the sheet on top of one another if there is more than one.