Fix common crash for large playbacks.
NSOperation seems to have troubles when chaining together thousands of
references. An operation sometimes returns a crash with a trace similar
to the following:

…
15  com.apple.CoreFoundation      	0x00007fff904804ef CFRelease + 591
16  com.apple.CoreFoundation      	0x00007fff90493749 -[__NSArrayM
dealloc] + 185
17  com.apple.Foundation          	0x00007fff932eadae
-[__NSOperationInternal dealloc] + 41
18  com.apple.Foundation          	0x00007fff932ead4f -[NSOperation
dealloc] + 55
19  com.apple.CoreFoundation      	0x00007fff904804ef CFRelease + 591
20  com.apple.CoreFoundation      	0x00007fff90493749 -[__NSArrayM
dealloc] + 185
21  com.apple.Foundation          	0x00007fff932eadae
-[__NSOperationInternal dealloc] + 41
22  com.apple.Foundation          	0x00007fff932ead4f -[NSOperation
dealloc] + 55
23  com.apple.CoreFoundation      	0x00007fff904804ef CFRelease + 591
24  com.apple.CoreFoundation      	0x00007fff90493749 -[__NSArrayM
dealloc] + 185
25  com.apple.Foundation          	0x00007fff932eadae
-[__NSOperationInternal dealloc] + 41
26  com.apple.Foundation          	0x00007fff932ead4f -[NSOperation
dealloc] + 55
…

While investigating the crash, the following Stack Overflow topic was
found:

http://stackoverflow.com/questions/19693079/nsoperationqueue-bug-with-de
pendencies

Based on the discussion of this topic, the behavior of Textual was
modified to dissolve the dependencies once an operation has concluded
running its assigned block rather than relying on NSOperationQueue to
do it for us.

The belief behind this change is that once the end of a dependency
chain is reached, all the dependencies are being released at once time
resulting in a loop such as the one seen above. Instead, we remove the
dependency when done with it to release it then.

It is unclear at this time if this solution is 100%. So far, there have
been no further crashes.