Add “finduser” command with optional input parameter.
If an input parameter is supplied, then the host mask of every user in the channel in which the command is performed is compared against the input. If the input is anywhere within the host mask, then the user is listed.
If no input parameter is supplied, then all users of the channel in which the command was performed is listed with their associated host masks.