WARNING: THIS COMMIT BREAKS MAVERICKS SUPPORT (for now)
Summary of commit:

To support both Yosemite and Mavericks, the main window was split into
two separate window files.

TXCMenuController.xib has two responsibilities now. It creates fresh
instances of TXMasterController, TXMenuController, and also sets up all
application menus.

TVCMainWindow*.xib is now responsible for the main window itself. This
file is loaded inside the -init of TXMasterController then the
-awakeFromNib method inside the master controller is still responsible
for app delegation.

This commit breaks Mavericks support because the main window file for
it is empty. That has not been converted over to the new format yet.

The focus of the commit was to finish work on the Yosemite version of
the text field. It now supports both dark and light variants.

For debugging purposes, there is code that can be commented at the top
of TVCMainWindowTextView.m in its -awakeFromNib method to switch
between dark and light mode.