Draw directly to screen instead of setting fill color.
Doing this allows the drawing to react to specific situations.

For example, when presenting a modal dialog, the fill color would not
update in time because by resignKey was called on the window delegate,
the main window -activeForDrawing did not see the modal which resulted
in it drawing as an active state.