Fixed topic drawing on Yosemite
When on Yosemite, the entire window of Textual is drawn into a layer
backed store which pretty much means the entire contents of the main
window is a giant image and only sections of it are redrawn when
necessary.

As part of this behavior change, the use of position: fixed; inside of
a style (mainly for the topic) will not work as expected when being
used in Textual on Yosemite because the WebView is part of the main
window.

To fix this, you can tell WebKit to use 3D graphics to update itself
correctly when inside a layer backed view which Textual is on Yosemite

To do this, all a style has to do is add the following CSS to elements
which are position: fixed

	-webkit-transform: translateZ(0);
	-webkit-font-smoothing: subpixel-antialiased;