temporary(?) fix for developer preview 5
Preview 5 of Yosemite appears to have changed the behavior of position:
fixed in WebView instances. At least in the context of Textual, the
WebView is now drawn as a layer backed view even though its never
explicitly defined as one neither is its host window.

As part of this, fonts do not look correct on position: fixed views
such as the topic. This fixes that.

Hopefully next developer preview will revert the layer backing.