Topic bar fix for bundled styles
The concept of the fix:

baseLayout.mustache now wraps the topic bar and content body in a
wrapper with the ID content_container. The content container block
houses the topic which is now positioned absolute relative to the top
of the view instead of using position: fixed which was the cause of the
abnormalities on Yosemite.

As part of this fix, Textual now handles scrolling itself. By this I
mean that when an element is appended, Textual will manually scroll the
view if it was already at the bottom before the append. The previous
behavior was to allow WebKit to handle this process but in order for
this fix to work, WebKit won’t be able to do it itself.

This commit is designed to be backwards compatible for the most part.
Hopefully styles were wise enough to target #body_home instead of
<body> when designing their layout so that the addition of
#body_container does not interfere with existing layouts.

A great deal of gratitude goes out towards @xlexi as he is the CSS
wizard that designed this workaround.