Automatically query apple.com for Mac model
Previously, Textual maintained a dictionary of all model identifiers
for the /sysinfo command and queried that whenever a user invoked the
command. However, Apple sometimes does not change the model identifier
when a refresh of a particular product line is not large enough to
justify it.

Therefore, to prevent false positives, the system profiler extension
will now query support-sp.apple.com when Textual is launched to obtain
the real model identifier based off the last 4 characters of the
current machines serial number.

As no personal information is transmitted during this transaction, this
behavior was not made optional. The model identifier found using this
method will be the preferred method and the dictionary stored within
Textual will be used as a fallback incase of failure.