Dictionary based template engine version
The key named “Template Engine Versions” is now read from
styleSettings.plist as a dictionary. The old “Template Engine Version”
value is completely ignored now.

The format of this dictionary is the version of Textual to target as
the key and the version of templates to target as the value.

For forward compatibility, a key named “default” can be specified which
will be used when a specific version is not targeted.