Add /autojoin/ command
Calling the /autojoin/ command will manually invoke auto join on
networks that do not have NickServ but are configured to wait for
responses from it.

This adds convenience when the command is added to the list of connect
commands.

The command itself will only perform auto join if it is not already in
progress or already happened.