Change build version scheme
The short build version is still maintained as 5.0.0, but going
forward, the long build version is now based on the date built. The
build date contains the month, day, and year with the hour built
separated by a period. iTunes Connect no longer allows the same version
number to be submitted twice. This commit is designed to workaround
that.