Wildcard URL matching
Instead of specifying each scheme that we support, we use a wildcard
based off the rules of RFC 3986.

The wildcard supports both two forward slashes as well as three.

In theory less matches should speed up the parser too, but I have not
benchmarked it. I do not see how this would be slower than the previous
implementation.