Updated user defaults assistant
Instead of writing to two objects depending on whether an object wants
to be written or read from the group container or sandbox container, we
instead ask +keyIsExcludedFromGroupContainer: for intent and go based
off that. It is slightly slower because it has to process the key, but
not enough for the human to notice.