Fix operator message logic
When sending or receiving messages targeted at a specific user mode
(e.g. “/msg +#channel” for voiced users), the logic responsible for
handling these messages did not take into account that a channel can
have the same prefix as a user mode.

This broke for a channel with a prefix such as + which resulted in the
+ symbol always being stripped because it was seen as a user mode.

The new logic will first look for whether there is a user mode prefix.
If it finds one, it checks the second character in the sequence. If the
second character is a channel prefix, then it truncates the first
character. Otherwise, the prefix is treated as part of the channel name.