Fix broken smiley converter plugin
The smiley converter plugin was recreating its database of emoticons
each time a message was received. This is because I removed an if
statement in -updateConversionTable in 5.0.1 to rebuild the database
each time the method was called instead of only doing it if
self.conversionTable was nil. However, I forgot that
-updateConversionTable was called when messages were rendered which led
to this overhead.