Fix mIRC color override for nox
The color-number attribute needs to be quoted for webkit (no,
really; verified with direct testing). Add an override for mIRC
color 6 (purple) to make it easier to read. Also remove the ltype
selector from the existing override, as it is entirely possible
for these colors to be used in ACTIONs as well as normal
PRIVMSGs.