Improve(?) menu validation in Textual
A submenu cannot validate normally because validation requires a target
an action. A dummy method, -emptyAction was added to the menu
controller. That way Textual can properly hide it when it wants during
validation.

Previously, the behavior was to manually hide the Channel Properties
submenu when a query is visible or the server console. Now, the submenu
is properly validated when it’s child menu becomes visible.

When switching selection, Textual will instruct the menu controller to
perform additional validation, but in a less hacked way than previously
used.

This change fixes Command I and other channel related keyboard
shortcuts not working because OS X did not recognize the keyboard
shortcuts until the menu with them became visible.