Let groundwork for Sparkle
What’s in this commit:

- Removed the TEXTUAL_BUILT_WITH_FORCED_BETA_LIFESPAN define and code
associated with it.
- Split the Developer ID configuration profile off into a “Base
Configuration” file which all others inherit from now instead of
inheriting from the Developer ID configuration profile.
- Changed it so that the Info.plist file is copied to the .tmp folder
and is modified there so that the Info.plist does not get accidentally
changed in a commit when the version information in it is dynamically
generated.
- Included the Sparkle framework, though not used yet. I