Do away with TPCPreferencesUserDefaults and TPCPreferencesUserDefaultsObjectProxy
TPCPreferencesUserDefaults and TPCPreferencesUserDefaultsObjectProxy
and have never worked as intended. This fixes the mess that they were.

TPCPreferencesUserDefaults now writes everything to the group
container. It no longer tries to make any balances between the group
container and app-level preference file.

TPCPreferencesUserDefaultsObjectProxy has been deprecated. It passes
responsibility on to TPCPreferencesUserDefaultsController.

TPCPreferencesUserDefaultsController is a subclass of
NSUserDefaultsController thats sole responsibility is to fire off
individual change notifications for iCloud. Adding a user defaults
controller to Interface Builder and using this class as the class of
the object is whats done now. It’s less hacked together and much
cleaner.

TPCPreferencesUserDefaultsController will always return the same
instance, even if you try to create a new one.

This commit also reverts 3b09c892396465e2bdb18bff4e49bdac73256a53,
bringing back the sandbox.

This commit also disables use of the Sparkle framework for now.

This commit also removes code for displaying a prompt for making
Textual the default IRC client. This code does not work within a
sandboxed environment.