Update TVCLogRenderer to allow formatting in and around URLs
The way in which TVCLogRenderer previously worked broke if a URL
contained any type of formatting (bold, italic, underline, color).
Therefore, when it encountered any formatting in a URL, it outright
stripped it. As a result, if a piece of text (for example, a topic) had
a background color, then the background color would apply until it got
to a URL, stop, then reappear after it.

To combat this, TVCLogRender no longer strips formatting from URLs.
However, there is a catch to this. If formatting existed inside of a
URL (“ex<bold start>ample.com”), then TVCLogRenderer would treat “ex”
and “ample.com” as separate URLs. To workaround this, TVCLogRenderer
now keeps track of when a URL begins (_rendererIsRenderingLinkIndex)
and uses that to lookup the correct URL from
TVCLogRendererResultsRangesOfAllLinksInBodyAttribute

“ex<bold start>ample.com” will continue to split up into separate links
to allow formatting, but both will now go to the correct destination
now (“example.com”)