Swizzle parts of NSUserDefaultsController
Out-of-the-box NSUserDefaults does not communicate changes directly
with NSUserDefaultsController. Therefore, when manually setting a
configuration option in Textual (maybe through setBool:forKey:), then
it would register the defaults in NSUserDefaults, but UI elements would
not update behind it.

TPCPreferencesUserDefaults now proxies all request to setObject:forKey:
to TPCPreferencesUserDefaultsController so the UI can update when
observed keys are changed. However, because TPCPreferencesUserDefaults
proxies setObject:forKey: elsewhere, if
TPCPreferencesUserDefaultsController tried to call it, which it would
under normal conditions, it would introduce a loop.

Therefore, I swizzled the method inside
TPCPreferencesUserDefaultsController responsible for setting values to
call a secondary method on TPCPreferencesUserDefaults. The logic is the
same in the swizzled method as can be seen by using Hopper
Disassembler.

Changes of rejection from store behind this change: 10x normal levels