Use NSObjectIsNotEmpty when referring to any keychain items
In version 5.0.3 of Textual, our keychain library returned nil for
keychain items that returned successfully but had a length of 0. After
updating Textual to use the keychain library packaged with Cocoa
Extensions, this is no longer the case. Therefore, when checking if a
value exists in the keychain, we must check its length rather than
whether it is nil or not.