Tracked down an impossible to find bug
For some reason, for certain outgoing messages, the control character
0x003 which defines colors were being inserted at the end of the
message. This broke certain comparisons that did an exact isEqual check
on the end of a message (for example, checking for “CBC” at the end).

Fixed the bug by specifying the proper variable to check in the if
statement whether this character should be included in a message.