Move away from hard coded user modes
IRCUser now defines the value -modes which is an NSString, ranked
highest to lowest, of all modes set on a user (e.g. q, a, o, h, v).
These modes are set based off NAMES & WHO replies as well as MODE
changes.

The -modes string is now used for determining “rank” - rank defines the
importance of one user over another. Rank is determined by the PREFIX=
value of the server configuration. The highest rank possible is 100,
which is the first mode in the PREFIX= value. It then counts backwards.
Textual sorts rank from highest to lowest.

By sorting by rank, instead of hard coded mode values, Textual can sort
properly for any possible combination of custom modes.

IRCUser -currentRank is now used throughout Textual in combination with
these changes. This method returns an enum value of IRCUserRank. This
value is an abstract representation of the privileges that the user
has. Instead of saying that UserX has mode +q, it returns
IRCUserChannelOwnerRank instead. By using this method, instead of
querying for the specific mode symbol, Textual is able to adopt for
different modes as IRC daemons change.

User coloring on the user list uses this very method for the purposes
of determining what color to use regardless of sort order or checking
the mode of the user directly.

To discourage use of checking modes directly in a plugin, the methods
-q, -a, -o, -h, and -v have all been deprecated in IRCUser to encourage
use of -currentRank