Add a historic_messages span to baseLayout.mustache
This span will host historic messages so that only a single opacity
declaration can be used, rather than one for each message.

This is a small change, so hopefully it will not severely break
existing styles, hence it not being a separate template version
altogether. If testing proves that it is a significant change, then it
will be made into a separate version in a later commit.