Introduce TXFormatDateTimeStringToCommonFormat
• Given string or date object, formats the value using NSDateFormatter
in long format style.
• Optionally will return original if the formatter returns a nil result.
• Returns the result in relative dates (e.g. “Yesterday”)

This commit is a fix for a server that sent timestamps in WHOWAS
replies in the following format: “Sun May 3 18:39:28 2015” — This
format is obviously strange with the time before the year.
NSDateFormatter was not able to parse this date, so now the original is
returned when it fails.