Refactor TVCQueuedCertificateTrustPanel
• Queued entries are now stored in an object
(TVCQueuedCertificateTrustPanelObject), rather than an array within an
array with fixed index values.
• A reference is no longer maintained to the actual trust. A reference
is maintained of the hosting socket and the callback block. When its
time to displayed a queued entry, then the socket is asked for the
trust at that point. This fixes a memory leak that would occur with
Textual maintaining a reference to the original trust.