Do not invoke -closeSocketWithError: for proxy populate errors
The socket hasn’t even attempted to connect at this point which means
if we close it, nothing will happen and the error message will just
linger in memory.

The error message that -socksProxyPopulateSystemSocksProxy: returns is
not critical which means we can simply log it to the console for later
review.