Add -socksProxyInUse to IRCConnectionSocket
-usesSocksProxy was being used to determine whether special actions
should be taken during read/write operations. -usesSocksProxy checks if
self.proxyType is equal to RCConnectionSocketSystemSocksProxyType and
if it was, it returns YES. Because its possible for self.proxyType to
equal IRCConnectionSocketSystemSocksProxyType but not actually being
using a proxy, Textual was getting confused. It was expecting to read
data from a proxy when no proxy was present.

For actions that require knowledge of a proxy, -socksProxyInUse is now
used because even if IRCConnectionSocketSystemSocksProxyType was the
original self.proxyType, it would be changed to SOCKS4 or SOCKS5 once
the system SOCKS proxy settings were successfully read.