Add option to enforce certain cipher suites
Added defaults key “Socket -> Secured Connection -> Enforce Allowed
Cipher Suites” which is set to NO by default that when turned on will
restrict access to the cipher suites defined by the defaults key
“Socket -> Secured Connection -> Allowed Cipher Suites”

The “Allowed Cipher Suites” defaults key contains a dictionary of
permitted cipher suites with the name of the cipher suite as the key
(unused, for reference only) and the integer of the cipher suite as its
value.

This feature is not designed to be user accessible. It exists more-so
for testing purposes.