Begin work on license system
This commits adds the public key used for verifying the signature of
files.

This commit also includes a function
(TLOLicenseManagerPublicKeyIsGenuine()) that can be used to check
whether the public key matches a known SHA-256 hash of the genuine
public key.

Commit also includes comment that details the goals of the license
system