Continued work on TLOLicenseManagerDownloader
All responses from the license API are now property lists, even on
success. Responses are expected to contain two required keys with one
optional.

‘Status Title’ is a string that will be presented to the end user as
the title of a dialog stating success or describing failure.
‘Status Message’ is a string that is presented to the end user as well
as part of a dialog providing additional information.

‘Status Context’ is an optional key that provides Textual context
information. This key is currently only used by license activation to
package the activated license data so that it can be written to disk.

The downloader tries very hard to ensure that the contents of each key
is of the expected type and provides logging for failures so that there
is no question where an event may fail.