Continued work on license manager
• Almost all TLOLicenseManager functions are now private, though the
header was already private so why does it matter?
• The method owned by TLOLicenseManagerDownloader for deactivating
license keys no longer accepts a license key as a parameter. Instead,
it will read the license key from TLOLicenseManager
• TLOLicenseManagerLicenseDictionary() now handles validation of the
user license file and will return nil if the signature is bad.
• TLOLicenseManagerUserLicenseWriteFileContents() now verifies the
input data is a property list and if the property list is a valid,
signed license. If it is valid, then it caches it right away.
• Added TLOLicenseManagerLicenseKeyIsValid() for verifying user entered
license keys.
• Removed several functions from TLOLicenseManager that are no longer
necessary.
• Replaces existing public key