Add Textual.dateChanged() JavaScript callback
To assist style authors with displaying date changes, Textual now has
built in system to watch for date changes (even compensating for DST)

This callback is invoked when two conditions are met:
1. The day has changed by reaching midnight.
2. The system clock has changed

For the first case (#1), the timer that handles the observation of
midnight understands Daylight Saving Time (DST) and other oddities.

For the second case (#2), Textual does not make an effort to compare if
the day has in-fact changed. It is merely passing the change down to
the style to inform it that the system clock changed.