Only present away messages every 5 minutes
When private messaging a user, if the user is away, an IRC server
typically sendings raw numeric 301 for every message received informing
the sender that the user is away.

This change adds the property -presentAwayMessageFor301 to IRCUser
which uses a timer to track whether the away message should be
presented. Under normal conditions, the away message is only shown once
every 5 minutes.