Large plugin infrastructure overhaul
• -didReceiveServerInputOnClient:senderInformation:messageInformation:
is now deprecated. This method will continue to be invoked on plugins
that support it, but it is recommended to eventually move to
-didReceiveServerInput:onClient:

-didReceiveServerInput:onClient: is passed an instance of
THOPluginDidReceiveServerInputConcreteObject which has properly cast
properties.

•
-didPostNewMessageForViewController:messageInfo:isThemeReload:isHistoryR
eload: is now deprecated. This method will continue to be invoked on
plugins that support it, but it is recommended to eventually move to
-didPostNewMessage:forViewController:

-didPostNewMessage:forViewController: is passed an instance of
THOPluginDidPostNewMessageConcreteObject which has properly cast
properties.

• The following methods which were deprecated in version 5.0.0 have
been removed and will no longer be invoked on loaded plugins:

-pluginLoadedIntoMemory:
-pluginUnloadedFromMemory

-pluginSupportsUserInputCommands
-pluginSupportsServerInputCommands

-preferencesView
-preferencesMenuItemName

-messageSentByUser:message:command:
-messageReceivedByServer:sender:message: