Backwards compatibility for -isSafeToPostNotification…
The last two commits introduced support for the BATCH command which can
be used to identify ZNC’s playback buffer in version 1.6 of ZNC and
later. This commit adds backwards compatibility for older versions by
using the logic that Textual 5.2.2 and earlier used when the BATCH
command is not available.